+++ CHANGELOG.rst
@@ -170,6 +170,8 @@ Changed
   `#334 <https://github.com/openego/eGon-data/issues/334>`_
 * Move functions used by multiple datasets
   `#323 <https://github.com/openego/eGon-data/issues/323>`_
+* Migrate scenario tables to datasets
+  `#309 <https://github.com/openego/eGon-data/issues/309>`_
 
 Bug fixes
 ---------